refactor: DTO 폴더 분리
- file과 folder의 DTO를 폴더로 분리했습니다.
