+++ src/main/java/com/c4cometrue/mystorage/controller/FileController.java
@@ -8,15 +8,18 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.c4cometrue.mystorage.dto.request.FileReq;
-import com.c4cometrue.mystorage.dto.request.UploadFileReq;
-import com.c4cometrue.mystorage.dto.response.FileDownloadRes;
-import com.c4cometrue.mystorage.dto.response.FileMetaDataRes;
+import com.c4cometrue.mystorage.dto.request.file.FileReq;
+import com.c4cometrue.mystorage.dto.request.file.MoveFileReq;
+import com.c4cometrue.mystorage.dto.request.file.UploadFileReq;
+import com.c4cometrue.mystorage.dto.response.file.FileDownloadRes;
+import com.c4cometrue.mystorage.dto.response.file.FileMetaDataRes;
 import com.c4cometrue.mystorage.service.FileService;
 
 import jakarta.validation.Valid;
@@ -34,11 +37,11 @@ public class FileController {
     /**
      * 파일 업로드 요청
      * @param req (파일, 사용자 이름, 폴더 기본키)
-     * @return {@link com.c4cometrue.mystorage.dto.response.FileMetaDataRes}
+     * @return {@link FileMetaDataRes}
      */
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
-    public FileMetaDataRes uploadFile(@Valid UploadFileReq req
+    public FileMetaDataRes uploadFile(@RequestBody @Valid UploadFileReq req
     ) {
         return fileService.uploadFile(req.file(), req.userName(), req.folderId());
     }
@@ -49,7 +52,7 @@ public class FileController {
      */
     @DeleteMapping
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteFile(@Valid FileReq req) {
+    public void deleteFile(@RequestBody @Valid FileReq req) {
         fileService.deleteFile(req.fileStorageName(), req.userName(), req.folderId());
     }
 
@@ -65,5 +68,10 @@ public class FileController {
             .contentType(MediaType.parseMediaType(file.mime()))
             .header(HttpHeaders.CONTENT_DISPOSITION, "attachment; filename=\"" + file.fileName() + "\"")
             .body(file.resource());
+    }
+
+    @PatchMapping
+    public void moveFile(@RequestBody @Valid MoveFileReq moveFileReq) {
+        // TODO : 파일 이동 구현
     }
 }
