+++ src/main/java/com/c4cometrue/mystorage/controller/FolderController.java
@@ -2,6 +2,7 @@ package com.c4cometrue.mystorage.controller;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.validation.annotation.Validated;
+import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -10,11 +11,13 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.c4cometrue.mystorage.dto.request.CreateFolderReq;
-import com.c4cometrue.mystorage.dto.request.GetFolderReq;
-import com.c4cometrue.mystorage.dto.request.UpdateFolderNameReq;
-import com.c4cometrue.mystorage.dto.response.CreateFolderRes;
-import com.c4cometrue.mystorage.dto.response.FolderOverviewRes;
+import com.c4cometrue.mystorage.dto.request.folder.CreateFolderReq;
+import com.c4cometrue.mystorage.dto.request.folder.DeleteFolderReq;
+import com.c4cometrue.mystorage.dto.request.folder.GetFolderReq;
+import com.c4cometrue.mystorage.dto.request.folder.MoveFolderReq;
+import com.c4cometrue.mystorage.dto.request.folder.UpdateFolderNameReq;
+import com.c4cometrue.mystorage.dto.response.folder.CreateFolderRes;
+import com.c4cometrue.mystorage.dto.response.folder.FolderOverviewRes;
 import com.c4cometrue.mystorage.service.FolderService;
 
 import jakarta.validation.Valid;
@@ -31,7 +34,7 @@ public class FolderController {
 	/**
 	 * 폴더의 개략적인 정보 요청
 	 * @param req (폴더 기본키, 폴더 이름, 사용자 이름, 부모 폴더 기본키)
-	 * @return {@link com.c4cometrue.mystorage.dto.response.FolderOverviewRes}
+	 * @return {@link FolderOverviewRes}
 	 */
 	@GetMapping
 	@ResponseStatus(HttpStatus.OK)
@@ -42,7 +45,7 @@ public class FolderController {
 	/**
 	 * 폴더 생성 요청이 성공하면 해당 폴더 pk를 포함한 정보 반환
 	 * @param req (폴더 이름, 사용자 이름, 부모 폴더 기본키)
-	 * @return {@link com.c4cometrue.mystorage.dto.response.CreateFolderRes}
+	 * @return {@link CreateFolderRes}
 	 */
 	@PostMapping
 	@ResponseStatus(HttpStatus.CREATED)
@@ -59,5 +62,17 @@ public class FolderController {
 	public void updateFolderName(@RequestBody @Valid UpdateFolderNameReq updateFolderNameReq) {
 		folderService.updateFolderName(updateFolderNameReq.folderId(), updateFolderNameReq.parentFolderId(),
 			updateFolderNameReq.userName(),	updateFolderNameReq.newFolderName());
+	}
+
+	@DeleteMapping
+	@ResponseStatus(HttpStatus.NO_CONTENT)
+	public void deleteFolder(@RequestBody @Valid DeleteFolderReq deleteFolderReq) {
+		// TODO: 폴더 삭제 구현
+	}
+
+	@PatchMapping
+	@ResponseStatus(HttpStatus.OK)
+	public void moveFolder(@RequestBody @Valid MoveFolderReq moveFolderReq) {
+		// TODO: 폴더 이동 구현
 	}
 }
