+++ src/main/java/com/c4cometrue/mystorage/dto/request/folder/CreateFolderReq.java
@@ -1,4 +1,4 @@
-package com.c4cometrue.mystorage.dto.request;
+package com.c4cometrue.mystorage.dto.request.folder;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
@@ -6,10 +6,9 @@ import jakarta.validation.constraints.NotNull;
 /**
  * @see com.c4cometrue.mystorage.entity.FolderMetaData
  */
-public record UpdateFolderNameReq(
-	@NotNull long folderId,
-	@NotNull long parentFolderId,
+public record CreateFolderReq(
+	@NotNull(message = "parent folder can't be null") long parentFolderId,
 	@NotBlank(message = "user name is blank") String userName,
-	@NotBlank(message = "new folder name is blank") String newFolderName
+	@NotBlank(message = "folder name is blank") String folderName
 ) {
 }
