+++ src/main/java/com/c4cometrue/mystorage/dto/request/folder/UpdateFolderNameReq.java
@@ -1,4 +1,4 @@
-package com.c4cometrue.mystorage.dto.request;
+package com.c4cometrue.mystorage.dto.request.folder;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
@@ -6,9 +6,10 @@ import jakarta.validation.constraints.NotNull;
 /**
  * @see com.c4cometrue.mystorage.entity.FolderMetaData
  */
-public record CreateFolderReq(
-	@NotNull(message = "parent folder can't be null") long parentFolderId,
+public record UpdateFolderNameReq(
+	@NotNull long folderId,
+	@NotNull long parentFolderId,
 	@NotBlank(message = "user name is blank") String userName,
-	@NotBlank(message = "folder name is blank") String folderName
+	@NotBlank(message = "new folder name is blank") String newFolderName
 ) {
 }
