+++ src/main/java/com/c4cometrue/mystorage/dto/response/file/FileMetaDataRes.java
@@ -1,4 +1,4 @@
-package com.c4cometrue.mystorage.dto.response;
+package com.c4cometrue.mystorage.dto.response.file;
 
 import com.c4cometrue.mystorage.entity.FileMetaData;
