+++ src/main/java/com/c4cometrue/mystorage/dto/response/folder/FolderMetaDataRes.java
@@ -1,4 +1,4 @@
-package com.c4cometrue.mystorage.dto.response;
+package com.c4cometrue.mystorage.dto.response.folder;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
