+++ src/test/java/com/c4cometrue/mystorage/controller/FileControllerTest.java
@@ -11,9 +11,9 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.core.io.Resource;
 
-import com.c4cometrue.mystorage.dto.request.FileReq;
-import com.c4cometrue.mystorage.dto.request.UploadFileReq;
-import com.c4cometrue.mystorage.dto.response.FileDownloadRes;
+import com.c4cometrue.mystorage.dto.request.file.FileReq;
+import com.c4cometrue.mystorage.dto.request.file.UploadFileReq;
+import com.c4cometrue.mystorage.dto.response.file.FileDownloadRes;
 import com.c4cometrue.mystorage.service.FileService;
 
 @ExtendWith(MockitoExtension.class)
